Adds the inital steps to gathering the commits and labels between two given tags
Adding in the calls to get and generate the logs for each valid commit

Working in promise hell

WIP

Fixing promise chain

Lint spaces fix

Adds logic to properly format and write to the changelog.md

Adds logic to properly format and write to the changelog.md

Adds the rest of the functionality to allow generation of the changelog. Adds the functionality to allow extras like pivotal links. Adds the abiliyt to have aliased labels and additional labels

Removes stuff